backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2
asyncpg==0.28.0
psycopg2-binary==2.9.9
